+++ .env.example
@@ -1,0 +1,10 @@
+# Google OAuth Configuration
+VITE_GOOGLE_CLIENT_ID=573120070871-0k7ga6ns79ie0jpg1ei6ip5vje2ostt6.apps.googleusercontent.com
+
+# Sui Configuration  
+VITE_SUI_FULLNODE_URL=https://fullnode.devnet.sui.io
+VITE_SUI_DEVNET_FAUCET=https://faucet.devnet.sui.io
+VITE_SUI_PROVER_DEV_ENDPOINT=https://prover-dev.mystenlabs.com/v1
+
+# Application Configuration
+VITE_REDIRECT_URI=https://sui-zklogin.vercel.app/
