+++ src/utils/constant.ts
@@ -1,22 +1,13 @@
-export const FULLNODE_URL = "https://fullnode.devnet.sui.io";
-
-export const CLIENT_ID =
+export const FULLNODE_URL = import.meta.env.VITE_SUI_FULLNODE_URL || "https://fullnode.devnet.sui.io";
+export const CLIENT_ID = import.meta.env.VITE_GOOGLE_CLIENT_ID || 
   "573120070871-0k7ga6ns79ie0jpg1ei6ip5vje2ostt6.apps.googleusercontent.com";
-
-export const REDIRECT_URI = "https://sui-zklogin.vercel.app/";
-
-export const SUI_DEVNET_FAUCET = "https://faucet.devnet.sui.io";
-
-export const SUI_PROVER_DEV_ENDPOINT = "https://prover-dev.mystenlabs.com/v1";
-
+export const REDIRECT_URI = import.meta.env.VITE_REDIRECT_URI || "https://sui-zklogin.vercel.app/";
+export const SUI_DEVNET_FAUCET = import.meta.env.VITE_SUI_DEVNET_FAUCET || "https://faucet.devnet.sui.io";
+export const SUI_PROVER_DEV_ENDPOINT = import.meta.env.VITE_SUI_PROVER_DEV_ENDPOINT || "https://prover-dev.mystenlabs.com/v1";
 export const KEY_PAIR_SESSION_STORAGE_KEY = "demo_ephemeral_key_pair";
-
 export const USER_SALT_LOCAL_STORAGE_KEY = "demo_user_salt_key_pair";
-
 export const RANDOMNESS_SESSION_STORAGE_KEY = "demo_randomness_key_pair";
-
 export const MAX_EPOCH_LOCAL_STORAGE_KEY = "demo_max_epoch_key_pair";
-
 export const STEPS_LABELS_TRANS_KEY = [
   "16e758e8",
   "9b8b5398",
@@ -26,7 +17,6 @@ export const STEPS_LABELS_TRANS_KEY = [
   "af802c7a",
   "c649dd70",
 ];
-
 export const STEPS_DESC = [
   "ephemeralKeyPair",
   "47b83f4e",
