+++ src/utils/theme/colors/success.ts
